bender: This introduces functionality to calculate the warpmap in both directions. This allows userspace locations to be mapped back to designspace locations. - new argument when calling the bender: bender(loc, space="user"). Values can be "design" (default): the location is in designspace, the result will be in user space. Value="user": the location is in userspace, the result will be in designspace. - validate warpmaps that they only increase. No decrease, no horizontals. - also build the reversed maps - new tests for the validation, and resutls
Note 1: this only works for maps that are tuples of in / out values. It does not provide a solution for maps defined by actual functions. I don't believe this is used widely so it might be possible to retire that functionality.
Note 2: this allows the bender to be used to calculate designspace coordinates for user space locations.